requirements.txt
Django==4.1.7
django-mysql==3.8.1
mysqlclient==2.1.1
django-extensions==3.2.1
djangorestframework==3.14.0
Pillow==9.0.0
requests==2.27.1
djangorestframework_simplejwt==4.7.2